feat(jules): implemented architectural hardening and CI/CD professionalization
- Implemented hard kill-switch based on ConvictionScore in `GovernanceGatekeeper.validate_inference`.
- Migrated W3C PROV-O loggers (`ProofOfThoughtLogger` and `MilestoneLogger`) from `adam_governance/state_control.py` into the centralized `src/governance/gatekeeper.py`.
- Configured `adam_governance/state_control.py` as a backward compatibility import proxy.
- Created root-level standard directories (`src/orchestrator/`, `src/agents/system_1/`, `src/agents/system_2/`, `src/core/`, `src/schemas/`, `tests/integration/`, `tests/stress/`) to align with professional-grade AI systems architecture.
- Added Diátaxis framework documentation folders (`docs/tutorials/`, `docs/how-to/`, `docs/explanation/`, `docs/reference/`).
- Authored institutional maturity policies (`SECURITY.md` and `CODE_OF_CONDUCT.md`).
- Formalized `daily_ritual.py` scheduling via a GitHub Action `.github/workflows/daily_ritual.yml`.
