+++ .github/workflows/daily_ritual.yml
@@ -43,6 +43,10 @@ jobs:
         run: |
           mkdocs build
 
+      - name: Run Provenance check
+        run: |
+          python ops/checks/check_provenance.py
+
       - name: Create Pull Request
         uses: peter-evans/create-pull-request@v6
         with:
