+++ SECURITY.md
@@ -1,10 +1,12 @@
 # Security Policy
 
 ## Supported Versions
-Only the latest version of AdamOS is actively supported with security updates.
+
+| Version | Supported          |
+| ------- | ------------------ |
+| 1.0.x   | :white_check_mark: |
+| < 1.0   | :x:                |
 
 ## Reporting a Vulnerability
-Please do not report security vulnerabilities through public GitHub issues.
 
-Instead, please report them directly via email to security@adamos.example.com.
-You should receive a response within 48 hours. If the issue is confirmed, we will release a patch as soon as possible depending on complexity.
+Please report vulnerabilities to the security team via email.
