+++ adam_governance/state_control.py
@@ -4,82 +4,4 @@ Dependencies: typing, hashlib, time
 Outputs: GovernanceGatekeeper base class.
 """
 
-from typing import Any, Dict
-import hashlib
-import time
-
-class GovernanceGatekeeper:
-    """
-    Deterministic gatekeeper that validates probabilistic inferences and payload provenance.
-    Enforces W3C PROV-O compliance on state transitions.
-    """
-
-    def entry_gate(self, payload: Dict[str, Any]) -> bool:
-        """Validate incoming payload structure and metadata before processing."""
-        if not payload.get("provenance"):
-            return False
-        return True
-
-    def exit_gate(self, result: Dict[str, Any]) -> Dict[str, Any]:
-        """Wrap outgoing results in an immutable ProvenanceHeader."""
-        content_hash = hashlib.sha256(str(result).encode('utf-8')).hexdigest()
-
-        provenance_header = {
-            "git_commit_hash": "placeholder_hash",
-            "timestamp": time.time(),
-            "content_hash": content_hash,
-            "jsonLogic_version": "1.0",
-        }
-
-        return {
-            "data": result,
-            "provenance": provenance_header
-        }
-
-class ProofOfThoughtLogger:
-    """
-    Enforces W3C PROV-O compliance strictly on every LLM-generated JSON payload.
-    """
-    def log_payload(self, payload: Dict[str, Any], derivation_path: str = "unknown", source_data_object: str = "unknown") -> Dict[str, Any]:
-        """Logs and structures a payload according to PROV-O."""
-        content_hash = hashlib.sha256(str(payload).encode('utf-8')).hexdigest()
-
-        prov_o_log = {
-            "entity": "LLM_Output",
-            "wasGeneratedBy": "Adam_Swarm_Agent",
-            "generatedAtTime": time.time(),
-            "content_hash": content_hash,
-            "payload": payload,
-            "derivation_path": derivation_path,
-            "source_data_object": source_data_object
-        }
-        return prov_o_log
-
-class MilestoneLogger:
-    """
-    Creates milestone markers and logs for tracking session state and runtime execution
-    downstream for human and machine learning reporting.
-    """
-    def __init__(self):
-        self.milestones: list = []
-
-    def add_milestone(self, name: str, details: Dict[str, Any], complexity: float, conviction: float) -> None:
-        """
-        Adds a milestone evaluating complexity and conviction.
-        """
-        self.milestones.append({
-            "name": name,
-            "details": details,
-            "complexity": complexity,
-            "conviction": conviction,
-            "timestamp": time.time()
-        })
-
-    def get_most_efficient_process(self) -> Dict[str, Any]:
-        """
-        Sorts to the most efficient deterministic process
-        (highest conviction over complexity).
-        """
-        if not self.milestones:
-            return {}
-        return sorted(self.milestones, key=lambda m: m["conviction"] / max(m["complexity"], 0.001), reverse=True)[0]
+from src.governance.gatekeeper import GovernanceGatekeeper, ProofOfThoughtLogger, MilestoneLogger
