+++ src/governance/gatekeeper.py
@@ -1,5 +1,6 @@
 import json
 import hashlib
+import time
 import urllib.request
 import urllib.error
 import asyncio
@@ -54,8 +55,8 @@ class GovernanceGatekeeper:
             raise GovernanceError(f"Invalid ProvenanceHeader: {e}")
 
         # Poison check based on confidence score boundary issues
-        if header.confidence_score < 0.0 or header.confidence_score > 1.0:
-             raise GovernanceError("Poisoned data detected: confidence score out of bounds.")
+        if header.confidence_score < 0.5 or header.confidence_score > 1.0:
+             raise GovernanceError("Poisoned data detected: confidence score out of bounds. Must be >= 0.5")
 
         # Extract the actual data payload to validate against the schema
         payload = inference_output.get("data", {})
@@ -151,3 +152,53 @@ class GovernanceGatekeeper:
             inference_output["observed_drift"] = False # Healed
 
         return self.validate_inference(inference_output)
+
+
+class ProofOfThoughtLogger:
+    """
+    Enforces W3C PROV-O compliance strictly on every LLM-generated JSON payload.
+    """
+    def log_payload(self, payload: Dict[str, Any], derivation_path: str = "unknown", source_data_object: str = "unknown") -> Dict[str, Any]:
+        """Logs and structures a payload according to PROV-O."""
+        content_hash = hashlib.sha256(str(payload).encode('utf-8')).hexdigest()
+
+        prov_o_log = {
+            "entity": "LLM_Output",
+            "wasGeneratedBy": "Adam_Swarm_Agent",
+            "generatedAtTime": time.time(),
+            "content_hash": content_hash,
+            "payload": payload,
+            "derivation_path": derivation_path,
+            "source_data_object": source_data_object
+        }
+        return prov_o_log
+
+
+class MilestoneLogger:
+    """
+    Creates milestone markers and logs for tracking session state and runtime execution
+    downstream for human and machine learning reporting.
+    """
+    def __init__(self):
+        self.milestones: list = []
+
+    def add_milestone(self, name: str, details: Dict[str, Any], complexity: float, conviction: float) -> None:
+        """
+        Adds a milestone evaluating complexity and conviction.
+        """
+        self.milestones.append({
+            "name": name,
+            "details": details,
+            "complexity": complexity,
+            "conviction": conviction,
+            "timestamp": time.time()
+        })
+
+    def get_most_efficient_process(self) -> Dict[str, Any]:
+        """
+        Sorts to the most efficient deterministic process
+        (highest conviction over complexity).
+        """
+        if not self.milestones:
+            return {}
+        return sorted(self.milestones, key=lambda m: m["conviction"] / max(m["complexity"], 0.001), reverse=True)[0]
